add light node client
