+++ cmd/client/cmd/flags.go
@@ -3,7 +3,7 @@
 *  @copyright defined in go-seele/LICENSE
  */
 
-package main
+package cmd
 
 import (
 	"github.com/seeleteam/go-seele/common"
