+++ cmd/client/light/main.go
@@ -1,0 +1,26 @@
+/**
+*  @file
+*  @copyright defined in go-seele/LICENSE
+ */
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/seeleteam/go-seele/cmd/client/cmd"
+	"github.com/urfave/cli"
+)
+
+func main() {
+	app := cli.NewApp()
+	app.Name = "light node client"
+	app.Usage = "interact with node process"
+	app.HideVersion = true
+
+	cmd.AddCommands(app, false)
+
+	if err := app.Run(os.Args); err != nil {
+		log.Fatal(err)
+	}
+}
